# document/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Content Manager Service",
        "contact": {},
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local testing"
        },
        {
            "url": "https://alinlandev.com/api/services/content",
            "description": "Server testing"
        }
    ],
    "paths": {
...
        "/categories/{id}/children": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Show Children",
                "operationId": "ShowChildren",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "actor",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    },
                    {
                        "name": "target-user",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false
            }
        },
        "/categories/{id}/nested": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Show nested",
                "operationId": "Shownested",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "actor",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    },
                    {
                        "name": "target-user",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false
            }
        },
        "/categories/{id}/products": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Products",
                "operationId": "Products",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false
            }
        },
...
        "/brands/{id}/products": {
            "get": {
                "tags": [
                    "Brand"
                ],
                "summary": "Products",
                "operationId": "GetProducts",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "headers": {}
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [],
...
}